+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.1.3
+
+Added ErrorToFailure mixin for easier error handling in repositories
+
 ## 0.1.2
 
 Added either_dart dependency instead of custom implementation
