+++ lib/q_architecture.dart
@@ -14,8 +14,9 @@ export 'src/domain/notifiers/simple_state_notifier.dart';
 export 'src/domain/providers/global_failure_provider.dart';
 export 'src/domain/providers/global_info_provider.dart';
 export 'src/domain/providers/global_loading_provider.dart';
-export 'src/domain/utils/int_extension.dart';
+export 'src/extensions/int_extension.dart';
 export 'src/presentation/widgets/paginated_list_view.dart';
 export 'src/presentation/widgets/base_widget.dart';
 export 'src/presentation/widgets/base_loading_indicator.dart';
 export 'src/data/mixins/error_to_failure_mixin.dart';
+export 'src/data/api_error_resolver.dart';
