+++ lib/src/extensions/iterable_extensions.dart
@@ -1,0 +1,8 @@
+extension IterableExtension<T> on Iterable<T> {
+  T? firstWhereOrNull(bool Function(T element) test) {
+    for (final element in this) {
+      if (test(element)) return element;
+    }
+    return null;
+  }
+}
